refactored forum subtopic page into its own component
